+++ pkg/sass_api/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add a `ReplaceSelectorVisitor` class.
 
+* `/` is now parsed as slash-separated lists rather than division in SassScript.
+
 ### Bogus Selectors
 
 * Drop support for bogus selectors that can never become valid CSS through