+++ pkg/sass_api/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 15.12.1
+
+* No user-visible changes.
+
 ## 15.12.0
 
 * No user-visible changes.